Add class "contains-task-list" to task lists
