+++ src/html/blocks/index.js
@@ -6,8 +6,7 @@ module.exports = [
     require('./blockquote'),
     require('./code'),
     require('./heading'),
-    require('./ulist'),
-    require('./olist'),
+    require('./list'),
     require('./listitem'),
     require('./unstyled'),
     table.table,
