+++ test/to-html/lists/ul-tasklist/output.html
@@ -1,4 +1,4 @@
-<ul>
+<ul class="contains-task-list">
 <li class="task-list-item"><input type="checkbox" class="task-list-item-checkbox" disabled /> Hello</li>
 <li class="task-list-item"><input type="checkbox" class="task-list-item-checkbox" checked disabled /> World</li>
 </ul>
